Add ability to fetch translated Loqate address
When fetching an address by ID, Loqate provides multiple
versions of this address in different languages, such as the
native language for the address and English.

It is sometimes required to have the English version of the
address, such as for FedEx shipments to China where FedEx only
supports ASCII characters.